requirements.txt
torch>=1.13.0
transformers>=4.28.0
datasets>=2.10.0
tabulate>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
scipy>=1.10.0
python-dotenv>=1.0.0
bitsandbytes>=0.37.0
peft>=0.4.0
tqdm>=4.65.0
plotly>=5.14.0
pillow>=9.5.0
seaborn>=0.12.0
shortuuid>=1.0.0